+++ SR/SR_full_dos.c
@@ -86,7 +86,7 @@ int SR_disassemble_convert_cjump(char *dst, const char *modifier, uint_fast32_t 
 {
     char cbuf[32];
     char *ostr, *str1;
-    int length1;
+    intptr_t length1;
 
     ostr = ud_insn_asm(&ud_obj);
 
@@ -128,7 +128,7 @@ int SR_disassemble_convert_fixup(const char *ostr, char *dst, fixup_data *fixup,
     int *label_value;
     output_data *output;
     uint_fast32_t sec, ofs;
-    int length1;
+    intptr_t length1;
 
     if (fixup->type == FT_SEGMENT)
     {
